refactor(main): import the wire decoders from the files that own them
Repoints the six importers at `remote-agent-decoding.ts`, `remote-device-decoding.ts` and
`remote-host-decoding.ts`, then deletes the manager's re-export block.

Six names stop being public in the process -- `decodeConversationReadState`,
`decodeConversationReadStates`, `decodeDuplicateBotResultFromHost`, `decodeRemoteDesktopCapabilities`,
`remoteAgentAvatarUrl`, `remoteServerLogoUrl`. Each is used only by the client that decodes with it;
none had an importer outside the manager. `isValidRemoteApiUrl` also stops being re-exported: it
belongs to `@openbot/contracts/invite-links`, and the manager was only forwarding it.

The `FromHost`/`FromMain` pointer in `src/preload/index.ts` moves in the same commit as the decoders
it points at. That convention is only useful while a reader can find the twin, so the two halves
cannot drift apart across commits.

Pure move.
