+++ src/main/index.ts
@@ -82,7 +82,8 @@ import { ProviderRuntimeManager } from "./provider-runtime-manager";
 import { RemoteDesktopManager } from "./remote-desktop-manager";
 import { resolveRemoteDesktopRuntime } from "./remote-desktop-runtime-artifact";
 import { loadOrCreateRemoteDesktopCredentials } from "./remote-desktop-secret-store";
-import { type DevelopmentRemoteServerConnection, decodeVoid, RemoteServerManager } from "./remote-server-manager";
+import { decodeVoid } from "./remote-host-decoding";
+import { type DevelopmentRemoteServerConnection, RemoteServerManager } from "./remote-server-manager";
 import { sendToRenderer } from "./renderer-ipc";
 import { canCheckRendererPermission, canRequestRendererPermission } from "./renderer-permissions";
 import { readSetupState, writeSetupState } from "./setup-store";
